Title: 가장 많이 겹치는 구간 2, Time: 17ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240830/가장 많이 겹치는 구간 2/section-with-maximum-overlap-2.cpp
@@ -10,7 +10,7 @@ int main(void) {
         checked[x1]++;
         checked[x2]--;
     }
-    int sum_val = 0;
+    long long sum_val = 0;
     int max = 1;
     for(long long ele : checked) {
         sum_val += ele;
